% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{fishatlas_figure_clusters}
\alias{fishatlas_figure_clusters}
\alias{fishatlas_figures}
\alias{fishatlas_studies}
\alias{fishatlas_cell_types}
\title{Get information about data on fishatlas.neuro.mpg.de}
\usage{

  fishatlas_figure_clusters(fishatlas_url = "https://fishatlas.neuro.mpg.de",
  study = "Kunst et al", figure = "Figure 3")

fishatlas_figures(fishatlas_url = "https://fishatlas.neuro.mpg.de",
  study = "Kunst et al")

fishatlas_studies(fishatlas_url = "https://fishatlas.neuro.mpg.de")

fishatlas_cell_types(fishatlas_url = "https://fishatlas.neuro.mpg.de",
  study = "Kunst et al")
}
\arguments{
\item{fishatlas_url}{the web location of the fish brain atlas project.}

\item{study}{the study you can to find information on. See all available using \code{fishatlas_studies()}}

\item{figure}{figure you want to find information on. See all available by giving \code{fishatlas_figures()} a study.}
}
\description{
Get information about data hosted by fishatlas.neuro.mpg.de, including contributing publications,
the figures for these publications that show data hosted by the site and the cell types that can be retreived from the site.
This information can be used with \code{fishatlas_read_neurons} to read specific neurons into R.
}
\examples{
\dontrun{
## What studies are available?
fishatlas_studies

## Huh, so what figures do we have in Kunst et al. 2019?
fishatlas_figures(study = "Kunst et al")

## Six figures. What neuron information did the last of them contain?
fishatlas_figure_clusters(study = "Kunst et al", figure = "Figure 6")

## What annotated cell types are there?
fishatlas_cell_types()

}
}
\references{
Kunst, Michael, Eva Laurell, Nouwar Mokayes, Anna Kramer, Fumi Kubo, António M. Fernandes, Dominique Förster, Marco Dal Maschio, and Herwig Baier. 2019. “A Cellular-Resolution Atlas of the Larval Zebrafish Brain.” Neuron, May. https://doi.org/10.1016/j.neuron.2019.04.034.
}
\seealso{
\code{\link{fishatlas_download_neurons}}, \code{\link{fishatlas_read_brain}}, \code{\link{fishatlas_read_saved_neurons}}
}
