% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meshes.R
\name{fishatlas_read_brain}
\alias{fishatlas_read_brain}
\title{Read 3D larval zebrafish template brain from Kunst et al. 2019}
\usage{
fishatlas_read_brain(type = c("brain_regions", "outline", "fiber",
  "cell_bodies"), fishatlas_url = "https://fishatlas.neuro.mpg.de",
  surf = NULL)
}
\arguments{
\item{type}{which version of the template brain you want to retreive. See details. Defaults to retreiving an
expert segmnetation of the template brain into 51 neuropils}

\item{fishatlas_url}{the web location of the fish brain atlas project}

\item{surf}{Use if you want to add the downloaded brain to a hxsruf object already in your environment. Defaults to null, a new object is made from scratch}
}
\value{
a \code{nat} package \code{\link[nat]{hxsurf}} object, which mimics the Amira surface format, replete with metadata that can be
accessed using \code{$}
}
\description{
Read the template brain for a 6 day old larval zebrafish, generated by Kunst et al. 2019,
straight from https://fishatlas.neuro.mpg.de.
3D triangular  meshes are returned as a \code{nat} package \code{\link[nat]{hxsurf}}
object, which mimics the Amira surface format. These can be be plotted in 3D using \code{rgl} and analysed with tools from the \code{nat} ecosystem.
This incldue subseting by neuropil, i.e.. if you only want to visualise or analyse the forebrain.
}
\details{
A single 3D brain object orits neuropils are read from .txt files hosted on https://fishatlas.neuro.mpg.de, and
turned into a \code{\link[nat]{hxsurf}} object in R, which mimics the Amira surface format.
As of May 2019, data available is:
\itemize{
 \item 	\emph{outline}	an outline from the zebrafish brain
 \item 	\emph{fiber}	 a 3D mesh based on a tract segmentation of the brain
 \item 	\emph{cell_bodies}	 a 3D mesh based on segmenting cell bodies of the brain
 \item 	\emph{brain_regions}	 51 expert annotated sub-regions of the brain. These include the major
 subdivisions beneth the forebrain, midbrain and hindbrain.
 Named in $Regionlist as 'major.neuropil_minor.neuropil_sub.region', e.g Forebrain_Diencephalon_Habenula
 }
 Note that since neurons are reconstructed from many different neural species,
 there is no 'standard' orientation between species, but within a species these neurons are registered to a
 template brain, usually using elastix.
}
\examples{
\dontrun{
## Lets have a look at these neuropils!
zfbrain = fishatlas_read_brain(type = "brain_regions")
plot3d(zfbrain)

## Maybe we just want to plot for forebrain
clear3d()
plot3d(subset(zfbrain, "Forebrain"), alpha = 0.5, col = "red")

## Maybe within the forebrain, we are interested in plotting the habenula
plot3d(subset(zfbrain, "Habenula"), alpha = 0.3, col = "red")

}
}
\references{
Kunst, Michael, Eva Laurell, Nouwar Mokayes, Anna Kramer, Fumi Kubo, António M. Fernandes, Dominique Förster, Marco Dal Maschio, and Herwig Baier. 2019. “A Cellular-Resolution Atlas of the Larval Zebrafish Brain.” Neuron, May. https://doi.org/10.1016/j.neuron.2019.04.034.
}
\seealso{
\code{\link{fishatlas_download_neurons}}, \code{\link{fishatlas_cell_types}}, \code{\link{fishatlas_read_saved_neurons}}
}
