% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{fishatlas_download_neurons}
\alias{fishatlas_download_neurons}
\alias{fishatlas_read_saved_neurons}
\alias{fishatlas_mirror_saved_neurons}
\title{Download all registered neuron data from the Fish Brain Atlas project}
\usage{

  fishatlas_download_neurons(fishatlas_url = "https://fishatlas.neuro.mpg.de",
  save.path = NULL)

fishatlas_read_saved_neurons(side = c("Right", "Left", "Original"),
  fishatlas_ids = NULL, save.path = NULL,
  fishatlas_url = "https://fishatlas.neuro.mpg.de")

fishatlas_mirror_saved_neurons(fishatlas_neurons, side = c("Right",
  "Left"), save.path = NULL,
  fishatlas_url = "https://fishatlas.neuro.mpg.de")
}
\arguments{
\item{fishatlas_url}{the web location of the fish brain atlas project}

\item{save.path}{if you do not want to save at \code{package_location/inst/exdata} then specify a location}

\item{side}{the side of the brain for which you want to read neurons. Neurons have been mirrored by Kunst et al. to have their
somata all on one side of the brain, or the other. The raw data, 'Original' can have somata at any location.}

\item{fishatlas_ids}{the name of fishatlas neurons to read or mirror, e.g. \code{T_161005_HuCxBG_erk_15_1}}

\item{fishatlas_neurons}{a \code{neuronlist} of neurons read using this package}
}
\description{
Download all neurons from the fishatlas.neuro.mpg.de.
These will be stored as a 53.7 MB at the location of this package, at inst/exdata or a user
specified location. These neuron data can be retreived from fishatlas.neuro.mpg.de in their original locations,
after registration to the template brain, or after having been mirrored to one hemisphere, either left or right.
We can therefore either read neurons for one hemisphere, or the other, or 'mirror' neurons using the saved data, rather than a saved
spatial transform. All neurons from Kunst et al. 2019.
}
\examples{
\dontrun{
## First we need to download all of the neurons
### We should only ever have to do this once!
fishatlas_download_neurons()

## Let's get all that sweet neuron data!
zfishn = fishatlas_read_saved_neurons(side = "Original")
plot3d(zfishn,soma = TRUE, lwd = 2)

## Hmm, but it would be better to have them all on the same side
clear3d()
zfishr = fishatlas_read_saved_neurons(side = "Right")
plot3d(zfishr,soma = TRUE, lwd = 2, col = "red")

## Great! How does that compare with neurons all on the left?
### Let's just look at the first 10 mirored.
zfishl = fishatlas_mirror_saved_neurons(fishatlas_neurons[1:10], side = "Left")
plot3d(zfishl,soma = TRUE, lwd = 2, col = "cyan")

}
}
\references{
Kunst, Michael, Eva Laurell, Nouwar Mokayes, Anna Kramer, Fumi Kubo, António M. Fernandes, Dominique Förster, Marco Dal Maschio, and Herwig Baier. 2019. “A Cellular-Resolution Atlas of the Larval Zebrafish Brain.” Neuron, May. https://doi.org/10.1016/j.neuron.2019.04.034.
}
\seealso{
\code{\link{fishatlas_cell_types}}, \code{\link{fishatlas_read_brain}}
}
